Remove RocksDbExtensions
